parse float

+++ requests/sabemiFunctions.js
@@ -133,11 +133,11 @@ module.exports.closeContract = async function closeContract(CodigoPessoaFisica,t
         "TabelaSelecionada": parseInt(tabela.codigoTabela),
         "ChaveSimulacao": simulationKey,
         "Prazo": parseInt(tabela.prazo.replace(",",".").replace(".","")),
-        "ValorVenda": parseInt(tabela.valorVenda.replace(",",".").replace(".","")),
-        "ValorLiquido": parseInt(tabela.valorLiquido.replace(",",".").replace(".","")),
-        "ValorParcela": parseInt(tabela.valorParcela.replace(",",".").replace(".","")),
-        "ValorPlano": parseInt(tabela.valorAP.replace(",",".").replace(".","")),
-        "ValorTaxa": parseInt(tabela.taxa.replace(",",".").replace(".","")),
+        "ValorVenda": parseFloat(tabela.valorVenda.replace(",",".").replace(".","")),
+        "ValorLiquido": parseFloat(tabela.valorLiquido.replace(",",".").replace(".","")),
+        "ValorParcela": parseFloat(tabela.valorParcela.replace(",",".").replace(".","")),
+        "ValorPlano": parseFloat(tabela.valorAP.replace(",",".").replace(".","")),
+        "ValorTaxa": parseFloat(tabela.taxa.replace(",",".").replace(".","")),
         "DataSimulacao": String(moment().toISOString())
       });
